Updates:  Enable additional warnings, fix a number of things.
None of them seem to have been genuine problems, but it's prettier now,
and some were questionable.

+++ guts/fts_open.c
@@ -4,7 +4,7 @@
  *	FTS * rc = NULL;
  */
  	char **rpath_argv;
-	size_t args = 0;
+	int args = 0;
 	int errored = 0;
 	int i;
